Update to refactor test base and interactions
